+++ ww/config_file.c
@@ -45,7 +45,7 @@ void unsafeCommitChanges(config_file_t *state)
     char     *string      = cJSON_PrintBuffered(state->root, (int) ((state->file_prebuffer_size) * 2), true);
     size_t    len         = strlen(string);
     const int max_retries = 3;
-    for (size_t i = 0; i < max_retries; i++)
+    for (int i = 0; i < max_retries; i++)
     {
         if (writeFile(state->file_path, string, len))
         {
